# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: isachore API
  description: |
    Chore management for households.

    ## Authenticating

    Two transports, either of which carries a session:

    - the httpOnly `isachore_token` cookie, which a browser sends on its own, and
    - the same opaque token as an `Authorization: Bearer` header, for clients with no cookie jar.

    Two ways to open one, and nothing downstream can tell them apart. `POST /api/v1/auth/login`
    takes an email and password, and answers `two_factor_required` instead of a session when the
    account has TOTP enabled - finish that with `POST /api/v1/auth/verify-2fa`. Where single
    sign-on is configured, `GET /api/v1/auth/oidc/start` is the other way in, and on a server
    with `OIDC_ONLY` set it is the only one: password login answers 403 there.
    `GET /api/v1/auth/methods` says which of the two this server offers, without a session.

    There is no self-registration: administrators create accounts, and the first administrator
    comes from the `init` CLI command.

    ## The CSRF header

    A request authenticated **by cookie** that uses an unsafe method (POST, PATCH, PUT, DELETE)
    must also carry a non-empty `X-CSRF-Token` header, or it is refused with 403 before it
    reaches the route. Any value will do: this is a custom-header defence over `SameSite=Lax`,
    so what matters is that a cross-site form cannot set the header at all.

...
  version: 0.1.0
paths:
...
  /api/v1/admin/settings:
    get:
      tags:
        - admin-settings
      summary: Read Settings
      operationId: read_settings_api_v1_admin_settings_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerSettingsRead'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Signed in, but not a site administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
      security:
        - sessionCookie: []
        - bearerToken: []
    patch:
      tags:
        - admin-settings
      summary: Update Settings
      operationId: update_settings_api_v1_admin_settings_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServerSettingsUpdate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerSettingsRead'
        '400':
          description: Email confirmation cannot be switched on while no SMTP server is configured, since nobody could then receive the link.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Signed in, but not a site administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - sessionCookie: []
        - bearerToken: []
  /api/v1/admin/settings/test-email:
    post:
      tags:
        - admin-settings
      summary: Send Test Email
      operationId: send_test_email_api_v1_admin_settings_test_email_post
      responses:
        '204':
          description: Successful Response
        '400':
          description: No SMTP server is configured, so there is nothing to test.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '401':
          description: 'No usable session: the cookie or bearer token is absent, expired, or belongs to a user who is no longer active.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Signed in, but not a site administrator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '429':
          description: Too many attempts. `Retry-After` carries the seconds still to run.
          headers:
            Retry-After:
              description: Seconds until the throttle window clears.
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '502':
          description: The SMTP server refused the message or could not be reached.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
      security:
        - sessionCookie: []
        - bearerToken: []
...
components:
  schemas:
...
    ErrorDetail:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
        - detail
      title: ErrorDetail
      description: |-
        The body FastAPI sends for a raised `HTTPException`.

        Declared so a refusal can be documented as carrying its explanation rather than as
        bodyless, which is what a route with only a `description` claims. Every hand-raised
        HTTPException in this app answers in this shape; only the routes that declare it say
        so, and wiring that up everywhere is the job of the security-scheme step in README's
        todo rather than of any single endpoint.
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ServerSettingsRead:
      properties:
        require_confirmation:
          type: boolean
          title: Require Confirmation
        smtp_configured:
          type: boolean
          title: Smtp Configured
        smtp_host:
          anyOf:
            - type: string
            - type: 'null'
          title: Smtp Host
        smtp_port:
          type: integer
          title: Smtp Port
        smtp_from:
          anyOf:
            - type: string
            - type: 'null'
          title: Smtp From
        oidc_configured:
          type: boolean
          title: Oidc Configured
        oidc_provider_name:
          type: string
          title: Oidc Provider Name
        oidc_issuer:
          anyOf:
            - type: string
            - type: 'null'
          title: Oidc Issuer
        oidc_client_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Oidc Client Id
        oidc_redirect_uri:
          type: string
          title: Oidc Redirect Uri
        oidc_only:
          type: boolean
          title: Oidc Only
      type: object
      required:
        - require_confirmation
        - smtp_configured
        - smtp_port
        - oidc_configured
        - oidc_provider_name
        - oidc_redirect_uri
        - oidc_only
      title: ServerSettingsRead
    ServerSettingsUpdate:
      properties:
        require_confirmation:
          type: boolean
          title: Require Confirmation
      type: object
      required:
        - require_confirmation
      title: ServerSettingsUpdate
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    sessionCookie:
      type: apiKey
      description: The httpOnly session cookie, set by `POST /api/v1/auth/login` or by the single sign-on callback. A browser sends it automatically; an unsafe method additionally needs the `X-CSRF-Token` header.
      in: cookie
      name: isachore_token
    bearerToken:
      type: http
      description: 'The same opaque session token as an `Authorization: Bearer` header, for clients with no cookie jar. Exempt from the `X-CSRF-Token` requirement.'
      scheme: bearer
    parkedAdminCookie:
      type: apiKey
      description: An impersonating administrator's own session, parked while they act as somebody else. `POST /api/v1/auth/stop-impersonating` is the only route it *authenticates* - several others read it to attribute an action to the real operator, and logout reads it to end both sessions.
      in: cookie
      name: isachore_admin_token